+++ src/PSRule.Rules.Azure/rules/Azure.Common.Rule.ps1
@@ -188,11 +188,26 @@ function global:IsWindowsOS {
     [OutputType([System.Boolean])]
     param ()
     process {
+        if ($Rule.TargetType -notin 'Microsoft.Compute/virtualMachines', 'Microsoft.Compute/virtualMachineScaleSets') {
+            return $False;
+        }
+        return ($TargetObject.Properties.storageProfile.osDisk.osType -eq 'Windows') -or
+            ($TargetObject.Properties.storageProfile.imageReference.publisher -in 'MicrosoftSQLServer', 'MicrosoftWindowsServer', 'MicrosoftVisualStudio', 'MicrosoftWindowsDesktop') -or
+            ($TargetObject.Properties.virtualMachineProfile.storageProfile.osDisk.osType -eq 'Windows') -or
+            ($TargetObject.Properties.virtualMachineProfile.storageProfile.imageReference.publisher -in 'MicrosoftSQLServer', 'MicrosoftWindowsServer', 'MicrosoftVisualStudio', 'MicrosoftWindowsDesktop')
+    }
+}
+
+function global:SupportsHybridUse {
+    [CmdletBinding()]
+    [OutputType([System.Boolean])]
+    param ()
+    process {
         if ($Rule.TargetType -ne 'Microsoft.Compute/virtualMachines') {
             return $False;
         }
         return ($TargetObject.Properties.storageProfile.osDisk.osType -eq 'Windows') -or
-            ($TargetObject.Properties.storageProfile.imageReference.publisher -in 'MicrosoftSQLServer', 'MicrosoftWindowsServer');
+            ($TargetObject.Properties.storageProfile.imageReference.publisher -in 'MicrosoftSQLServer', 'MicrosoftWindowsServer')
     }
 }